[NewIR]Move ir_parser.h into core/parser
