+++ paddle/pir/core/parser/ir_parser.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/pir/core/ir_parser.h"
+#include "paddle/pir/core/parser/ir_parser.h"
 
 #include "paddle/pir/core/builtin_dialect.h"
 #include "paddle/pir/core/builtin_type.h"
